feat: simplify sidebar container files
