--- components/sidebar/icons/FriendsIcon.tsx
@@ -1,63 +1,0 @@
-const DEFAULT_SIZE = 28;
-
-const FriendsIcon = ({
-  selected,
-  width = DEFAULT_SIZE,
-  height = DEFAULT_SIZE,
-  onClick,
-}: {
-  width?: number;
-  height?: number;
-  selected?: boolean;
-  onClick: () => void;
-}) => (
-  <button type="button" onClick={onClick}>
-    <svg
-      width={width}
-      height={height}
-      viewBox="0 0 24 24"
-      fill="none"
-      className={`stroke-current${selected ? " fill-current" : ""}`}
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M7 20C7 18.3431 9.23858 17 12 17C14.7614 17 17 18.3431 17 20"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M18 14.2495C19.7659 14.7124 21 15.7697 21 16.9999"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M6 14.2495C4.2341 14.7124 3 15.7697 3 16.9999"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M12 14C13.6569 14 15 12.6569 15 11C15 9.34315 13.6569 8 12 8C10.3431 8 9 9.34315 9 11C9 12.6569 10.3431 14 12 14Z"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M18 10.2361C18.6137 9.68679 19 8.8885 19 8C19 6.34315 17.6569 5 16 5C15.2316 5 14.5308 5.28885 14 5.76389"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M6 10.2361C5.38625 9.68679 5 8.8885 5 8C5 6.34315 6.34315 5 8 5C8.76835 5 9.46924 5.28885 10 5.76389"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  </button>
-);
-
-export default FriendsIcon;
