+++ components/sidebar/types.tsx
@@ -1,0 +1,1 @@
+export type SidebarOptions = "friends" | "notifications" | "playback" | "none";
